# openapi.yaml
openapi: 3.0.3
info:
  title: Dropbox Media Server API
  description: API for shared media gallery, public file requests, and administration.
  version: 1.6.0
servers:
  - url: /
paths:
...
  /api/share/{hash}/files:
    get:
      summary: List files in a public share
      parameters:
        - $ref: "#/components/parameters/hash"
        - name: recursive
          in: query
          schema:
            type: string
            enum: ["0", "1"]
      responses:
        "200":
          description: A list of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryFile"

  # Public Share Download
  /api/share/{hash}/file/{path}:
    get:
      summary: Get a file from a public share (redirects)
      parameters:
        - $ref: "#/components/parameters/hash"
        - $ref: "#/components/parameters/path"
        - name: download
          in: query
          schema:
            type: string
            enum: ["1"]
        - name: inline
          in: query
          schema:
            type: string
            enum: ["true"]
      responses:
        "302":
          description: Redirect to FileBrowser download URL
  /api/share/{hash}/download:
    get:
      summary: Download all files in a share as a ZIP
      parameters:
        - $ref: "#/components/parameters/hash"
      responses:
        "200":
          description: ZIP file stream

  # Public Media Utilities
...
  /api/analytics/shares:
    get:
      summary: Get summary stats for all shares
      security:
        - BearerAuth: []
      parameters:
        - name: include_empty
          in: query
          schema: { type: boolean, default: true }
      responses:
        "200":
          description: Analytics data
  /api/analytics/shares/{hash}:
    get:
      summary: Get detailed stats for a specific share
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/hash"
      responses:
        "200":
          description: Detailed stats
...
components:
  parameters:
    hash:
      name: hash
      in: path
      required: true
      schema: { type: string }
    path:
      name: path
      in: path
      required: true
      schema: { type: string }
  securitySchemes:
    XAuth:
      type: apiKey
      in: header
      name: X-Auth
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    GalleryFile:
      type: object
      properties:
        name: { type: string }
        path: { type: string }
        type: { type: string, enum: [image, video, file] }
        size: { type: integer }
        extension: { type: string }
        inline_url: { type: string }
        download_url: { type: string }
...